dev/us-east-1/mwaa/requirements.txt
aws-cdk-lib==2.96.2
constructs>=10.0.0,<11.0.0
PyYAML
